package/requirements.txt
lxml>=4.6.1
absl-py>=0.10.0
matplotlib>=3.0.3
numpy>=1.19.4
Pillow>=6.0.0
PyYAML>=5.1
six>=1.15.0
tensorflow>=2.4.0
tensorflow-addons>=0.12
tensorflow-hub>=0.11
neural-structured-learning>=1.3.1
tensorflow-model-optimization>=0.5
Cython>=0.29.13
tensorflow-datasets
gin-config 
pycocotools
py-cpuinfo
gputil
opencv-python
wandb
wget
